+++ frontend/src/components/browser/main.vue
@@ -93,10 +93,13 @@ export default {
 
 <style scoped lang="scss">
 @use "vuetify/styles/settings/variables" as vuetify;
-@import "../book-cover.scss";
+@use "sass:map";
+@use "../book-cover" as bookcover;
+
 $top-toolbar-margin: 94px;
 $card-margin: 32px;
 $browse-pane-margin-top: calc($top-toolbar-margin + $card-margin);
+
 #browsePane {
   margin-top: $browse-pane-margin-top;
   margin-left: max($card-margin, env(safe-area-inset-left));
@@ -104,19 +107,23 @@ $browse-pane-margin-top: calc($top-toolbar-margin + $card-margin);
   margin-bottom: max($card-margin, env(safe-area-inset-bottom));
   overflow: auto;
 }
+
 .browsePaneSearch {
   margin-top: calc($browse-pane-margin-top + 32px) !important;
 }
+
 #browsePaneContainer {
   margin-top: $card-margin;
   display: grid;
-  grid-template-columns: repeat(auto-fit, $cover-width);
+  grid-template-columns: repeat(auto-fit, bookcover.$cover-width);
   grid-gap: $card-margin;
   align-content: flex-start;
 }
+
 .padFooter {
   padding-bottom: 45px !important;
 }
+
 .placeholder {
   position: fixed;
   height: 50vh !important;
@@ -125,6 +132,7 @@ $browse-pane-margin-top: calc($top-toolbar-margin + $card-margin);
   left: 50%;
   transform: translate(-50%, -50%);
 }
+
 #searchLimitMessage {
   padding-top: 20px;
   text-align: center;
@@ -132,8 +140,9 @@ $browse-pane-margin-top: calc($top-toolbar-margin + $card-margin);
   color: rgb(var(--v-theme-textDisabled));
 }
 
-@media #{map-get(vuetify.$display-breakpoints, 'sm-and-down')} {
+@media #{map.get(vuetify.$display-breakpoints, 'sm-and-down')} {
   $small-card-margin: 16px;
+
   #browsePane {
     margin-left: max($small-card-margin, env(safe-area-inset-left));
     margin-right: max($small-card-margin, env(safe-area-inset-right));
@@ -141,7 +150,7 @@ $browse-pane-margin-top: calc($top-toolbar-margin + $card-margin);
   }
 
   #browsePaneContainer {
-    grid-template-columns: repeat(auto-fit, $small-cover-width);
+    grid-template-columns: repeat(auto-fit, bookcover.$small-cover-width);
     grid-gap: $small-card-margin;
     justify-content: space-evenly;
   }