+++ frontend/src/components/browser/toolbars/browser-toolbars-head.vue
@@ -26,6 +26,8 @@ export default {
 
 <style scoped lang="scss">
 @use "vuetify/styles/settings/variables" as vuetify;
+@use "sass:map";
+
 #codexHeadToolbarWrapper {
   z-index: 100;
   position: fixed;
@@ -37,7 +39,7 @@ export default {
   padding-right: calc(10px + env(safe-area-inset-right) / 4);
 }
 
-@media #{map-get(vuetify.$display-breakpoints, 'sm-and-down')} {
+@media #{map.get(vuetify.$display-breakpoints, 'sm-and-down')} {
   .codexHeadToolbar {
     padding-top: env(safe-area-inset-top);
     padding-left: calc(5px + env(safe-area-inset-left) / 4);