+++ frontend/src/components/reader/book-change-activator.vue
@@ -51,7 +51,7 @@ export default {
 };
 </script>
 <style scoped lang="scss">
-@import "./change-column.scss";
+@forward "./change-column";
 
 .leftPos {
   left: 0px;